feat: support replace with stage attachment
