+++ tests/data/sample_3_duplicate.csv
@@ -1,0 +1,8 @@
+1,'Beijing',100
+2,'Shanghai',80
+3,'Guangzhou',60
+4,'Shenzhen',70
+5,'Shenzhen',55
+6,'Beijing',99
+10,'Hong Kong‘,88
+10,'Suzhou‘,66
