+++ tests/data/sample_3_replace.csv
@@ -1,0 +1,5 @@
+1,'Chengdu',80
+3,'Chongqing',90
+6,'HangZhou',92
+9,'Changsha',91
+10,'Hong Kong‘,88
