+++ tests/suites/1_stateful/07_stage_attachment/07_0000_insert_with_stage.sh
