+++ tests/suites/1_stateful/07_stage_attachment/07_0001_replace_with_stage.result
@@ -1,0 +1,49 @@
+sample.csv
+96
+227
+null
+1	'Beijing'	100	China
+2	'Shanghai'	80	China
+3	'Guangzhou'	60	China
+4	'Shenzhen'	70	China
+5	'Shenzhen'	55	China
+6	'Beijing'	99	China
+96
+null
+1	'Beijing'	100	China
+2	'Shanghai'	80	China
+3	'Guangzhou'	60	China
+4	'Shenzhen'	70	China
+5	'Shenzhen'	55	China
+6	'Beijing'	99	China
+77
+null
+1	'Beijing'	2	China
+2	'Shanghai'	2	China
+3	'Guangzhou'	2	China
+4	'Shenzhen'	2	China
+5	'Shenzhen'	2	China
+6	'Beijing'	2	China
+83
+null
+1	'Chengdu'	80	China
+2	'Shanghai'	2	China
+3	'Chongqing'	90	China
+4	'Shenzhen'	2	China
+5	'Shenzhen'	2	China
+6	'HangZhou'	92	China
+9	'Changsha'	91	China
+10	'Hong Kong‘	88	China
+133
+{
+  "code": 4000,
+  "message": "duplicated data detected in merge source"
+}
+1	'Chengdu'	80	China
+2	'Shanghai'	2	China
+3	'Chongqing'	90	China
+4	'Shenzhen'	2	China
+5	'Shenzhen'	2	China
+6	'HangZhou'	92	China
+9	'Changsha'	91	China
+10	'Hong Kong‘	88	China
